# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS «Медицина Долголетия» — API
  version: "1.0.0"
  description: |
    API платформы обучения врачей. Авторизация — через httpOnly cookie `token` (JWT),
    выдаётся эндпоинтами `/api/auth/login` и `/api/auth/register`.
    Роли: `student` (врач), `curator`, `admin`, `super_admin`.
servers:
  - url: /api
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [student, curator, admin, super_admin] }
        specialization: { type: string, nullable: true }
        workplace: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        stream_id: { type: string, nullable: true }
        referral_code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Регистрация врача (или назначение роли из приглашения по email)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string }
                password: { type: string, minLength: 6 }
                name: { type: string }
                specialization: { type: string, description: "Обязательно для роли student" }
                phone: { type: string }
                workplace: { type: string }
                ref: { type: string, description: "Реферальный код пригласившего (необязательно)" }
      responses:
        "200": { description: "Успешная регистрация, выдана cookie token" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: "Email уже зарегистрирован" }
        "429": { $ref: "#/components/responses/TooManyAttempts" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Вход по email и паролю
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: "Успешный вход, выдана cookie token" }
        "401": { description: "Неверный email или пароль" }
        "429": { $ref: "#/components/responses/TooManyAttempts" }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Текущий пользователь
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { user: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Смена пароля (нужен текущий пароль)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 6 }
      responses:
        "200": { description: OK }
        "401": { description: "Текущий пароль неверен" }
...
  /streams:
    get:
      tags: [Calendar]
      summary: Список потоков
      responses: { "200": { description: OK } }
    post:
      tags: [Calendar]
      summary: Создать поток
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string }, startDate: { type: string, format: date } }
      responses: { "200": { description: OK } }
  /streams/{id}:
    delete:
      tags: [Calendar]
      summary: Удалить поток
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /events:
    get:
      tags: [Calendar]
      summary: Список прямых эфиров
      responses: { "200": { description: OK } }
    post:
      tags: [Calendar]
      summary: Создать эфир (можно с еженедельным повтором)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, date, time]
              properties:
                title: { type: string }
                date: { type: string, format: date }
                time: { type: string, example: "18:00" }
                duration: { type: integer, example: 60 }
                speaker: { type: string }
                streamId: { type: string, nullable: true, description: "null = виден всем потокам" }
                joinUrl: { type: string }
                description: { type: string }
                recurrence: { type: string, enum: [weekly], description: "Необязательно — еженедельный повтор" }
                recurrenceUntil: { type: string, format: date }
      responses: { "200": { description: OK } }
  /events/{id}:
    delete:
      tags: [Calendar]
      summary: Удалить эфир (?series=true — удалить всю серию повторов)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: series, in: query, schema: { type: boolean } }
      responses: { "200": { description: OK } }
...
components:
  responses:
    Unauthorized:
      description: "Не авторизован или сессия отозвана"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: "Некорректные входные данные"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyAttempts:
      description: "Слишком много попыток — подождите несколько минут"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }